Further loader code consolidation
moved _utils.py to loader/_utils.py
